refactor: move notification position enum
