+++ src/app/common/components/notification/Notification.stories.tsx
@@ -1,7 +1,8 @@
 import { Story, Meta } from '@storybook/react';
 import { StorybookVariants } from 'test/storybook.helper';
 import { Button } from '../button/Button';
-import { Notification, NotificationPosition, NotificationProps } from './Notification';
+import { Notification, NotificationProps } from './Notification';
+import { NotificationPosition } from './notification.types';
 import { NotificationProvider, useNotification } from './NotificationProvider';
 
 export default {
