+++ src/app/common/components/notification/notification.types.ts
@@ -1,7 +1,15 @@
 import { ReactNode } from 'react';
-import { NotificationPosition } from './Notification';
 
 export type NotificationId = number | string;
+
+export enum NotificationPosition {
+  Top = 'top',
+  TopLeft = 'top-left',
+  TopRight = 'top-right',
+  Bottom = 'bottom',
+  BottomLeft = 'bottom-left',
+  BottomRight = 'bottom-right',
+}
 
 export interface NotificationConfig {
   id: NotificationId;
